Refactor complex components to reduce code complexity
Break down large components into smaller, more maintainable pieces:

- Extract Home page sections into separate components
- Split CharacterShowcase into reusable CharacterCard component
- Modularize FormShowcase with individual field components
- Separate ThemeShowcase into themed color sections
- Extract InteractiveShowcase dialog and dropdown components
- Refactor UserService getUsers method with query helpers

Resolves Codacy complexity warnings while maintaining functionality.
